refactor: 인수 테스트에서의 `RANDOM_PORT` 사용을 위한 템플릿 작성
