+++ src/test/java/roomescape/acceptance/AdminTest.java
@@ -1,12 +1,10 @@
-package roomescape;
+package roomescape.acceptance;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-public class AdminTest {
+class AdminTest extends AcceptanceTest {
 
     @Test
     @DisplayName("홈 화면을 요청하면 200 OK을 응답한다.")
